🍎 LazyColumn Scroll 관리 기능 추가
## 🍎 LazyColumn Scroll 관리 기능 추가
### 🍀 1. FloatingActionButton 생성
- Scaffold 에 FloatingActionButton 추가

### 🍀 2. LazyColumn 의 스크롤이 최상단이 아닌 경우 FloatingActionButton 보이게 설정
- `rememberLazyListState` 를 사용
  - firstVisibleItemIndex : 처음 보이는 Item index 를 가져옴
  - isScrollInProgress : 스크롤 중인지 여부
  - animateScrollToItem : index 값을 넣으면 해당 index 값으로 Scroll 움직임
  - stopScroll : 스크롤 중지
